Relabel bulletformat (#79)
Elo   | -0.00 +- 8.02 (95%)
Conf  | 4.0+0.04s Threads=1 Hash=8MB
Games | N: 2004 W: 484 L: 484 D: 1036
Penta | [13, 224, 530, 220, 15]```
https://chess.aronpetkovski.com/test/7583/

No I will not run a sprt

bench: 2914563

+++ src/defs.h
@@ -19,6 +19,7 @@
 #include <sstream>
 #include <memory>
 #include <iomanip>
+#include <fstream>
 
 __extension__ typedef unsigned __int128 uint128_t;
 
@@ -38,7 +39,7 @@ inline void set_bit(uint64_t &bitboard, uint8_t square)
 {
     bitboard |= (1ULL << (square));
 }
-inline uint8_t get_bit(uint64_t bitboard, uint8_t square)
+inline uint64_t get_bit(uint64_t bitboard, uint8_t square)
 {
     return (bitboard & (1ULL << (square)));
 }
